+++ CHANGELOG.md
@@ -10,7 +10,11 @@ _This project follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- **Configuration:** Added support to configure IBKR broker
+
 ### Changed
+
+- **Configuration:** Different UI/UX improvements
 
 ### Fixed